pyproject.toml
[project]
name = "karo-localization-llm"
version = "1.0.0"
description = "LLM-powered localization for .xcstrings files using Google Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.3",
]
